1.0 beta

+++ Kamienica/src/main/java/kamienica/conventer/ReadingEnergyConverter.java
@@ -1,12 +1,22 @@
 package kamienica.conventer;
 
+
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.stereotype.Component;
 
-import kamienica.model.ReadingEnergy;
-import kamienica.service.ReadingService;
+import kamienica.feature.reading.ReadingEnergy;
+import kamienica.feature.reading.ReadingService;
 
+/**
+ * 
+ * @author macfol
+ * 
+ *         The class's role is to convert string value from the jsp form into
+ *         object from DB
+ *
+ */
 @Component
 public class ReadingEnergyConverter implements Converter<Object, ReadingEnergy>{
 
@@ -15,7 +25,7 @@ public class ReadingEnergyConverter implements Converter<Object, ReadingEnergy>{
 	
 	@Override
 	public ReadingEnergy convert(Object element) {
-		Integer id = Integer.parseInt((String)element);
+		Long id = Long.parseLong((String) element);
 		return readingService.getEnergyById(id);
 	}
