+++ Kamienica/src/main/java/kamienica/conventer/ReadingWaterConverter.java
@@ -4,9 +4,16 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.stereotype.Component;
 
-import kamienica.model.ReadingWater;
-import kamienica.service.ReadingService;
-
+import kamienica.feature.reading.ReadingService;
+import kamienica.feature.reading.ReadingWater;
+/**
+ * 
+ * @author macfol
+ * 
+ *         The class's role is to convert string value from the jsp form into
+ *         object from DB
+ *
+ */
 @Component
 public class ReadingWaterConverter implements Converter<Object, ReadingWater> {
 
@@ -15,7 +22,7 @@ public class ReadingWaterConverter implements Converter<Object, ReadingWater> {
 
 	@Override
 	public ReadingWater convert(Object element) {
-		Integer id = Integer.parseInt((String) element);
+		Long id = Long.parseLong((String) element);
 		return readingService.getWaterById(id);
 	}
